`applyCover` and `applyBounty` should be allowed when paused
As when we have a potential hack event, we shall pause the pool first while investigating, so that no new insurance will be sold during the time we undertaking the investigation.

Therefore, when applyCover() and applyBounty(), most likely the pool will be in paused status.

And we should allow it rather than require the paused status to be lifted first.